pyproject.toml
[project]
name = "news-report-agent"
version = "0.1.0"
description = "热点资讯聚合 Agentic AI - 基于 LangGraph 和 DeepAgents 的智能新闻分析系统"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    # 核心框架
    "deepagents>=0.2.8",
    "langgraph>=0.2.0",
    "langgraph-supervisor>=0.0.31",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.1.0",
    # 搜索与数据获取
    "tavily-python>=0.7.13",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    # 数据处理
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    # 持久化
    "langgraph-checkpoint-sqlite>=0.1.0",
    # 调度
    "apscheduler>=3.10.0",
    # 工具类
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
news-agent = "cli.main:main"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["."]
